requirements.txt
pygame
pycryptodome
numpy
